refactor: Move ServiceDiscovery to src/Host/ServiceDiscovery
