+++ src/Host/ServiceDiscovery/ServiceDiscoveryOptions.cs
@@ -1,4 +1,4 @@
-namespace Host;
+namespace Host.ServiceDiscovery;
 
 /// <summary>
 /// Configuration options for automatic service discovery and DI registration.
@@ -26,7 +26,7 @@ public sealed class ServiceDiscoveryOptions
     /// Specific types to ignore during service registration.
     /// </summary>
     public List<Type> ExcludedTypes { get; set; } = [];
-    
+
     public List<string> AllowedLayers { get; set; } = [];
 
     /// <summary>
@@ -43,7 +43,7 @@ public sealed class ServiceDiscoveryOptions
     /// If false, ServiceDiscovery will not log debug messages.
     /// </summary>
     public bool EnableLogging { get; set; } = true;
-    
+
     /// <summary>
     /// Default service lifetime for all discovered types.
     /// </summary>
